hello-game-app/hello-frontend/requirements.txt
google-auth==2.43.0
google-cloud-pubsub==2.33.0
google-api-core==2.28.1
gunicorn==23.0.0
flask==3.1.2
requests==2.32.5